Began to serialize Pickables for level loading. LevelStatus holds a list of Pickables.
Moved SaveAndLoad to Player-directory

Changed namespace to correlated scripts
