+++ Assets/Scripts/Pickables/Pickables.cs
@@ -5,6 +5,7 @@ using Unity.Mathematics;
 using UnityEngine;
 using UnityEngine.Experimental.Rendering.Universal;
 
+[Serializable]
 public class Pickables : MonoBehaviour
 {
     [SerializeField]
